rename folder main to app
