+++ src/app/pages/terms-and-conditions.jsx
